third_party/WebKit/LayoutTests/reporting-observer/reporting-api.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/WebKit/public/platform/reporting.mojom.js"></script>
<script>
// Mock implementation of ReportingServiceProxy.
// |promise| property is always a promise for the next report to be queued.
class MockReportingServiceProxy {
  constructor() {
    this.bindingSet = new mojo.BindingSet(blink.mojom.ReportingServiceProxy);
    this.resetPromise();
  }

  bind(handle) {
    this.bindingSet.addBinding(this, handle);
  }

  resetPromise() {
    this.promise = new Promise((resolve, reject) => {
      this.resolve = resolve;
    });
  }

  // Interface implementation.
  async queueDeprecationReport(url, message, sourceFile, lineNumber) {
    this.resolve([url, message, sourceFile, lineNumber]);
    this.resetPromise();
  }
}

// Make an instance and have it receive the request.
var proxy = new MockReportingServiceProxy();
var interceptor = new MojoInterfaceInterceptor(blink.mojom.ReportingServiceProxy.name);
interceptor.oninterfacerequest = e => proxy.bind(e.handle);
interceptor.start();

promise_test(async () => {
  let promise = proxy.promise;

  // Use a deprecated feature.
  window.webkitStorageInfo;

  // Ensure the report is generated and routed to the mojo interface.
  let [url, message, sourceFile, lineNumber] = await promise;
  assert_true(url.url.endsWith("reporting-observer/reporting-api.html"));
  assert_equals(typeof message, "string");
  assert_true(sourceFile.endsWith("reporting-observer/reporting-api.html"));
  assert_equals(typeof lineNumber, "number");
}, "Deprecation report");
</script>
